redis-mysql-py/requirements.txt
mysql-connector-python==8.3.0
redis==5.2.1
timeit_decorator==2.0.0